backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
requests==2.31.0
chardet==5.2.0
PyPDF2==3.0.1
aiofiles==23.2.1 